pyproject.toml
[project]
name = "npl_mcp"
version = "0.1.0"
description = "Simple FastMCP hello‑world server and NPL MCP package"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.0.0,<4.0.0",
    "pyyaml>=6.0",
    "pillow>=10.0",
    "aiosqlite>=0.19.0",
    "asyncpg>=0.29.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "playwright>=1.40.0",
    "sse-starlette>=1.6.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=5.0",
    "html2text>=2024.1.0",
    "pdfplumber>=0.10.0",
    "lxml>=5.0.0",
    "shortuuid>=1.0.0",
    "cairosvg>=2.8.2",
    "anthropic>=0.79.0",
    "openai>=2.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-playwright>=0.4.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src/npl_mcp"]
omit = [
    "*/__main__.py",
    "*/__init__.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/npl_mcp", "tools"]

[project.scripts]
npl-mcp = "npl_mcp.launcher:main"
npl-docs-regen = "npl_mcp.docs_regen:main"
npl-tmlanguage = "npl_mcp.scripts.tmlanguage:main"
git-dump = "tools.git_dump:main"
git-tree = "tools.git_tree:main"
2md = "tools.convert_to_markdown:main"
md-view = "tools.md_view:main"
view-md = "tools.view_md:main"
